# packages/ivr-migration-api/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IVR Migration API",
    "description": "Manages creation of IVR migration resources.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "def-13": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Task resource id.",
            "type": "string"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "in-progress",
              "success",
              "error"
            ],
            "description": "Task execution status"
          },
          "testExecutionId": {
            "description": "Unique id of test execution",
            "type": "string"
          },
          "createdAt": {
            "description": "Date/time created",
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "description": "ID of owner.",
            "type": "string"
          },
          "statusMessage": {
            "description": "message for the status",
            "type": "string"
          },
          "progress": {
            "description": "total progress of the task",
            "type": "number"
          },
          "itemsTotal": {
            "description": "total number of items in the task",
            "type": "number"
          },
          "itemsSucceeded": {
            "description": "total number of items succeeded",
            "type": "number"
          },
          "itemsFailed": {
            "description": "no. of items failed",
            "type": "number"
          },
          "itemsCompleted": {
            "description": "no. of items completed",
            "type": "number"
          },
          "updatedAt": {
            "description": "Date/time updated",
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "description": "Last ID of user who made a change.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "taskStatus",
          "testExecutionId",
          "createdAt",
          "createdBy",
          "itemsTotal",
          "itemsSucceeded",
          "itemsFailed",
          "itemsCompleted"
        ],
        "title": "taskResource"
      },
      "def-14": {
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/def-13"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "lastEvaluated": {
                "description": "Last evaluated task Id",
                "type": "string"
              }
            }
          }
        },
        "required": [
          "tasks"
        ],
        "title": "executionTasks_list"
      },
      "def-15": {
        "type": "object",
        "properties": {
          "taskId": {
            "description": "Recommendation task id.",
            "type": "string"
          },
          "taskItemId": {
            "description": "Recommendation task item id.",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "description": "Execution task item state"
          },
          "statusMessage": {
            "description": "failure message",
            "type": "string"
          },
          "recommendationDownloadUrl": {
            "description": "recommendation download url",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "taskItemId"
        ],
        "title": "taskItemResource"
      },
      "def-16": {
        "type": "object",
        "properties": {
          "taskItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/def-15"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "count": {
                "type": "number"
              },
              "token": {
                "description": "Recommendation task item id.",
                "type": "string"
              }
            }
          }
        },
        "required": [
          "taskItems"
        ],
        "title": "taskItemListResource"
      }
    }
  },
  "paths": {
...
    "/testExecutions/{testExecutionId}/recommendationTasks": {
      "get": {
        "tags": [
          "Execution Tasks"
        ],
        "description": "Lists Execution Tasks",
        "parameters": [
          {
            "schema": {
              "default": 20,
              "type": "integer"
            },
            "in": "query",
            "name": "count",
            "required": true,
            "description": "Count of results to return, if not specified default to 20"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "fromTaskId",
            "required": false,
            "description": "Token used to paginate from (exclusive)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "testExecutionId",
            "required": true,
            "description": "Unique id of test execution"
          }
        ],
        "x-security-scopes": [
          "admin",
          "contributor",
          "reader"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "$ref": "#/components/schemas/def-14"
                },
                "examples": {
                  "List of Execution Tasks": {
                    "summary": "Paginated list of executionTasks",
                    "value": {
                      "tasks": [
                        {
                          "id": "XP8QITQ9OR",
                          "taskStatus": "success",
                          "itemsTotal": 11,
                          "itemsSucceeded": 11,
                          "itemsFailed": 0,
                          "itemsCompleted": 11,
                          "createdAt": "2024-10-30T03:30:58.734Z",
                          "createdBy": "someone@somewhere.com",
                          "testExecutionId": "01jbdqm00vdyz1d2bdbp26qm9p"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/recommendationTasks/{taskId}/taskItems": {
      "get": {
        "operationId": "listExecutionTaskItems",
        "tags": [
          "Task Items"
        ],
        "description": "Lists execution task items",
        "parameters": [
          {
            "schema": {
              "default": 20,
              "type": "integer"
            },
            "in": "query",
            "name": "count",
            "required": true,
            "description": "Count of results to return, if not specified default to 20"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "fromDate",
            "required": false,
            "description": "Id to paginate from (exclusive)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "taskId",
            "required": true,
            "description": "Recommendation task id."
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin",
          "contributor",
          "reader"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "$ref": "#/components/schemas/def-16"
                },
                "examples": {
                  "List of task items": {
                    "summary": "Paginated list of execution task items",
                    "value": {
                      "taskItems": [
                        {
                          "taskId": "01jbdv60dzd26j43wvn762ehm4",
                          "taskItemId": "01jb8zyt5ccs2zrq5c2vsmrx7a",
                          "status": "success",
                          "statusMessage": "Essential container in task exited"
                        },
                        {
                          "taskId": "01jbdv60dzd26j43wvn762ehm4",
                          "taskItemId": "01jb8zyt5ccs2zrq5c2vsmrx7a",
                          "status": "success",
                          "statusMessage": "Essential container in task exited"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:30001"
    }
  ],
  "security": [],
...
}